+++ packages/react-querybuilder/src/types/type-fest/index.ts
@@ -1,4 +1,4 @@
-// Version 4.40.1
+// Version 4.41.0
 export * from './and';
 export * from './except';
 export * from './greater-than';