+++ packages/react-querybuilder/src/utils/formatQuery/__tests__/dbquery.drizzle.postgres.test.ts
@@ -4,9 +4,9 @@ import { PGlite } from '@electric-sql/pglite';
 import { boolean, integer, pgTable, text } from 'drizzle-orm/pg-core';
 import { drizzle } from 'drizzle-orm/pglite';
 import { convertToIC, formatQuery } from 'react-querybuilder';
-import type { TestSQLParams } from './drizzleTestUtils';
-import { dbSetup, dbTestsDrizzle, superUsers } from './drizzleTestUtils';
-import { generateDrizzleRuleGroupProcessor } from './generateDrizzleRuleGroupProcessor';
+import type { TestSQLParams } from '../dbqueryTestUtils';
+import { dbSetup, superUsers } from '../dbqueryTestUtils';
+import { dbTestsDrizzle } from '../dbqueryDrizzleTestUtils';
 
 const columnsPostgres = {
   firstName: text().notNull(),
@@ -17,8 +17,8 @@ const columnsPostgres = {
   powerUpAge: integer(),
 };
 
-const tablePostgres = pgTable('superusers', columnsPostgres);
-const superusers = superUsers('postgres');
+const superusers = pgTable('superusers', columnsPostgres);
+const superUsersPostgres = superUsers('postgres');
 
 const pglitePostgresDB = new PGlite();
 const drizzlePostgresDB = drizzle({ client: pglitePostgresDB, schema: { superusers } });
@@ -26,16 +26,13 @@ const drizzlePostgresDB = drizzle({ client: pglitePostgresDB, schema: { superuse
 const testPostgres = ({ query, expectedResult, fqOptions }: TestSQLParams) => {
   test.each(['standard', 'independent combinators'])('%s', async testType => {
     const queryToTest = testType === 'independent combinators' ? convertToIC(query) : query;
-    const sql = formatQuery(queryToTest, {
-      ...fqOptions,
-      ruleGroupProcessor: generateDrizzleRuleGroupProcessor(tablePostgres),
-    });
-    const result = await drizzlePostgresDB.select().from(tablePostgres).where(sql);
+    const where = formatQuery(queryToTest, { ...fqOptions, format: 'drizzle' });
+    const result = await drizzlePostgresDB.query.superusers.findMany({ where });
     expect(result).toEqual(expectedResult);
   });
 };
 
-describe('Drizzle (PostgreSQL)', () => {
+describe('Drizzle relational queries (PostgreSQL)', () => {
   beforeAll(async () => {
     await pglitePostgresDB.exec(dbSetup('postgres'));
   });
@@ -46,16 +43,16 @@ describe('Drizzle (PostgreSQL)', () => {
 
   // Test for coverage
   test('undefined column', async () => {
-    const sql = formatQuery(
+    const where = formatQuery(
       { combinator: 'and', rules: [{ field: 'doesNotExist', operator: '=', value: '' }] },
-      { ruleGroupProcessor: generateDrizzleRuleGroupProcessor(tablePostgres) }
+      { format: 'drizzle' }
     );
-    const result = await drizzlePostgresDB.select().from(tablePostgres).where(sql);
-    expect(result).toEqual(superusers);
+    const result = await drizzlePostgresDB.query.superusers.findMany({ where });
+    expect(result).toEqual(superUsersPostgres);
   });
 
   // Common tests
-  for (const [name, t] of Object.entries(dbTestsDrizzle(superusers))) {
+  for (const [name, t] of Object.entries(dbTestsDrizzle(superUsersPostgres))) {
     describe(name, () => {
       testPostgres(t);
     });